requirements.txt
ace_tools==0.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
numpy==2.2.2
pandas==2.2.3
playwright==1.50.0
psycopg2-binary==2.9.10
pyee==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
